Tabular_to_Neo4j/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.10
openai>=1.0.0
requests>=2.25.0
python-dotenv>=0.19.0
